fix permissions to controllers

+++ controllers/GroupController.php
@@ -168,7 +168,11 @@ class GroupController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+		if($id > 2){
+			$this->findModel($id)->delete();
+		} else {
+            throw new NotFoundHttpException(Yii::t('app', 'This item can not be delete.'));
+        }
 
         return $this->redirect(['index']);
     }
@@ -185,7 +189,7 @@ class GroupController extends Controller
         if (($model = Group::findOne($id)) !== null) {
             return $model;
         } else {
-            throw new NotFoundHttpException('The requested page does not exist.');
+            throw new NotFoundHttpException(Yii::t('app', 'The requested page does not exist.'));
         }
     }
 }
